Update on front end
- Using inheritance
