+++ ConsoleApplication1/ConsoleApplication1/ConsoleApplication1.vcxproj
@@ -144,7 +144,7 @@
     <ClCompile Include="Character.cpp" />
     <ClCompile Include="ConsoleApplication1.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Settings.cpp" />
+    <ClCompile Include="Screen.cpp" />
     <ClCompile Include="Textbox.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -154,7 +154,6 @@
     <ClInclude Include="main.h" />
     <ClInclude Include="obstacle.h" />
     <ClInclude Include="Screen.h" />
-    <ClInclude Include="Settings.h" />
     <ClInclude Include="Textbox.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
