+++ ConsoleApplication1/ConsoleApplication1/x64/Debug/ConsoleA.3e1a4322.tlog/ConsoleApplication1.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.35.32215:TargetPlatformVersion=10.0.22000.0:
-Debug|x64|D:\PROJECT CS202\CS202-CrossRoad-Game\ConsoleApplication1\|
+Debug|x64|C:\Users\Admin\source\repos\2ND\CS202-CrossRoad-Game\ConsoleApplication1\|
