+++ ConsoleApplication1/ConsoleApplication1/x64/Debug/ConsoleApplication1.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>D:\PROJECT CS202\CS202-CrossRoad-Game\ConsoleApplication1\x64\Debug\ConsoleApplication1.exe</FullPath>
+      <FullPath>C:\Users\Admin\source\repos\2ND\CS202-CrossRoad-Game\ConsoleApplication1\x64\Debug\ConsoleApplication1.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
