+++ ConsoleApplication1/ConsoleApplication1/x64/Debug/ConsoleApplication1.log
@@ -1,2 +1,4 @@
 ﻿  ConsoleApplication1.cpp
-  ConsoleApplication1.vcxproj -> D:\PROJECT CS202\CS202-CrossRoad-Game\ConsoleApplication1\x64\Debug\ConsoleApplication1.exe
+  main.cpp
+  Generating Code...
+  ConsoleApplication1.vcxproj -> C:\Users\Admin\source\repos\2ND\CS202-CrossRoad-Game\ConsoleApplication1\x64\Debug\ConsoleApplication1.exe
